+++ numpy/core/src/common/cblasfuncs.c
@@ -9,6 +9,7 @@
 #include <Python.h>
 
 #include "numpy/arrayobject.h"
+#include "numpy/npy_math.h"
 #include "npy_cblas.h"
 #include "arraytypes.h"
 #include "common.h"
