+++ numpy/core/src/npysort/npysort_common.h
@@ -3,6 +3,7 @@
 
 #include <stdlib.h>
 #include <numpy/ndarraytypes.h>
+#include <numpy/npy_math.h>
 
 #ifdef __cplusplus
 extern "C" {
