+++ numpy/core/src/umath/scalarmath.c.src
@@ -16,6 +16,7 @@
 #include "numpy/arrayobject.h"
 #include "numpy/ufuncobject.h"
 #include "numpy/arrayscalars.h"
+#include "numpy/npy_math.h"
 
 #include "npy_import.h"
 #include "npy_pycompat.h"
